+++ lib/src/ast/sass/configured_variable.dart
@@ -5,14 +5,16 @@
 import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
+import '../../util/span.dart';
 import 'expression.dart';
+import 'declaration.dart';
 import 'node.dart';
 
 /// A variable configured by a `with` clause in a `@use` or `@forward` rule.
 ///
 /// {@category AST}
 @sealed
-class ConfiguredVariable implements SassNode {
+class ConfiguredVariable implements SassNode, SassDeclaration {
   /// The name of the variable being configured.
   final String name;
 
@@ -25,6 +27,8 @@ class ConfiguredVariable implements SassNode {
   final bool isGuarded;
 
   final FileSpan span;
+
+  FileSpan get nameSpan => span.initialIdentifier(includeLeading: 1);
 
   ConfiguredVariable(this.name, this.expression, this.span,
       {bool guarded = false})