+++ lib/src/parse/css.dart
@@ -128,7 +128,7 @@ class CssParser extends ScssParser {
           "This function isn't allowed in plain CSS.", scanner.spanFrom(start));
     }
 
-    return FunctionExpression(
+    return InterpolatedFunctionExpression(
         // Create a fake interpolation to force the function to be interpreted
         // as plain CSS, rather than calling a user-defined function.
         Interpolation([StringExpression(identifier)], identifier.span),