+++ lib/src/value/color.dart
@@ -305,3 +305,25 @@ class SassColor extends Value {
     return buffer.toString();
   }
 }
+
+/// Extension methods that are only visible through the `sass_api` package.
+///
+/// These methods are considered less general-purpose and more liable to change
+/// than the main [SassColor] interface.
+extension SassApiColor on SassColor {
+  /// Whether the `red`, `green`, and `blue` fields have already been computed
+  /// for this value.
+  ///
+  /// Note that these fields can always be safely computed after the fact; this
+  /// just allows users such as the Sass embedded compiler to access whichever
+  /// representation is readily available.
+  bool get hasCalculatedRgb => _red != null;
+
+  /// Whether the `hue`, `saturation`, and `lightness` fields have already been
+  /// computed for this value.
+  ///
+  /// Note that these fields can always be safely computed after the fact; this
+  /// just allows users such as the Sass embedded compiler to access whichever
+  /// representation is readily available.
+  bool get hasCalculatedHsl => _saturation != null;
+}